+++ instrumentation/opentelemetry-instrumentation-urllib/tests/test_urllib_integration.py
@@ -137,8 +137,8 @@ class RequestsIntegrationTestBase(abc.ABC):
                 self.assertGreaterEqual(view_data.aggregator.current.sum, 0)
 
     def test_name_callback(self):
-        def name_callback():
-            return "test_name"
+        def name_callback(method, url):
+            return "GET" + url
 
         URLLibInstrumentor().uninstrument()
         URLLibInstrumentor().instrument(name_callback=name_callback)
@@ -147,10 +147,10 @@ class RequestsIntegrationTestBase(abc.ABC):
         self.assertEqual(result.read(), b"Hello!")
         span = self.assert_span()
 
-        self.assertEqual(span.name, "test_name")
+        self.assertEqual(span.name, "GET" + self.URL)
 
     def test_name_callback_default(self):
-        def name_callback():
+        def name_callback(method, url):
             return 123
 
         URLLibInstrumentor().uninstrument()